Clean up some IDG history.
Change moxa-sp7 to moxa-sm2, and got rid of all traces of
"sp7". Basically, whenever the main lab cryostat changes, the name of
the moxa should be changed to the expected one for that cryostat

Removed the for-LAM.conf file, as the devices are now at LAM

Post-Murdock shift of the Optics Lab hosts to standard PFS identities.